Move fs and signal handling to internal pkg
